test: Compare integers by same signs
`BOOST_*` allows for comparison of integers with different signs. This
is problematic in the case of overflow and underflow, and should not be
valid for a unit test framework.

+++ src/test/dbwrapper_tests.cpp
@@ -363,11 +363,11 @@ BOOST_AUTO_TEST_CASE(iterator_ordering)
                 break;
             BOOST_CHECK(it->GetKey(key));
             if (x & 1) {
-                BOOST_CHECK_EQUAL(key, x + 1);
+                BOOST_CHECK_EQUAL(static_cast<unsigned int>(key), x + 1);
                 continue;
             }
             BOOST_CHECK(it->GetValue(value));
-            BOOST_CHECK_EQUAL(key, x);
+            BOOST_CHECK_EQUAL(static_cast<unsigned int>(key), x);
             BOOST_CHECK_EQUAL(value, x*x);
             it->Next();
         }
